+++ R/eovcheck.R
@@ -332,15 +332,10 @@ eovcheck_ggplot2 = function(diagnosticInfo, xlab, ylab, col, smoother, twosd) {
     aes(x = .data$fitted, y = .data$residuals)
   ) +
     geom_point(shape = 1) +
-    geom_hline(yintercept = 0, linetype = 3, colour = "lightgrey") +
+    geom_hline(yintercept = 0, linetype = 3, colour = "lightgrey", linewidth = 0.4) +
     labs(x = xlab, y = ylab, title = "") +
     coord_cartesian(ylim = yRange) +
-    theme(
-      panel.background = element_rect(fill = "white", colour = NA),
-      panel.grid.major = element_blank(),
-      panel.grid.minor = element_blank(),
-      panel.border = element_rect(fill = NA, colour = "black")
-    )
+    s20x_ggplot2_base_theme()
 
   if (smoother) {
     smootherColour = if (is.null(col)) {
@@ -348,7 +343,12 @@ eovcheck_ggplot2 = function(diagnosticInfo, xlab, ylab, col, smoother, twosd) {
     } else {
       col
     }
-    plotObject = plotObject + geom_smooth(method = "loess", se = FALSE, colour = smootherColour)
+    plotObject = plotObject + geom_smooth(
+      method = "loess",
+      se = FALSE,
+      colour = smootherColour,
+      linewidth = 0.5
+    )
   }
 
   if (twosd) {
@@ -356,7 +356,7 @@ eovcheck_ggplot2 = function(diagnosticInfo, xlab, ylab, col, smoother, twosd) {
       yintercept = twoSdValues,
       linetype = 3,
       colour = "grey",
-      linewidth = 2
+      linewidth = 0.5
     )
   }